src/pyproject.toml
[tool.poetry]
name = "bewise-test"
version = "0.1.0"
description = ""
authors = ["deniko1985 <deniko1985@gmail.com>"]
readme = "README.md"
# packages = [{include = "bewise_test"}]

[tool.poetry.dependencies]
python = "3.10.6"
alembic = "1.11.0"
anyio = "3.6.2"
asyncpg = "0.27.0"
certifi = "2023.5.7"
charset-normalizer = "3.1.0"
click = "8.1.3"
databases = "0.7.0"
fastapi = "0.95.1"
greenlet = "2.0.2"
h11 = "0.14.0"
idna = "3.4"
Jinja2 = "3.1.2"
Mako = "1.2.4"
MarkupSafe = "2.1.2"
psycopg2-binary = "2.9.6"
pydantic = "1.10.7"
pydub = "0.25.1"
python-dotenv = "1.0.0"
python-multipart = "0.0.6"
requests = "2.30.0"
sniffio = "1.3.0"
SQLAlchemy = "1.4.48"
starlette = "0.26.1"
typing_extensions = "4.5.0"
urllib3 = "2.0.2"
uvicorn = "0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
